pyproject.toml
[project]
name = "gallery_inspector"
version = "1.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "imageio~=2.37.0",
    "loguru~=0.7.3",
    "matplotlib==3.10.8",
    "openpyxl~=3.1.5",
    "pandas~=2.2.3",
    "pillow==12.1.0",
    "pymediainfo~=7.0.1",
    "rawpy==0.26.0",
    "tqdm==4.67.2",
    "streamlit==1.53.1",
    "streamlit-aggrid==1.2.1.post2",
    "streamlit-extras>=0.7.8",
    "plotly==6.5.2",
    "customtkinter==5.2.2",
    "pandas-stubs~=2.2.3",
    "tkinterdnd2-universal>=1.7.3",
    "ruff>=0.14.14",
    "black>=26.1.0",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.18.0",
]
